Remove the last bit of DBus
