+++ pepper2/common/drive_types/no_action.py
@@ -7,7 +7,7 @@ from pepper2.common.constraint import Constraint, TrueConstraint
 from .drive_type import DriveType
 
 if TYPE_CHECKING:
-    from pepper2.daemon.dbus.drive import Drive
+    from pepper2.common.drive import Drive
 
 
 class NoActionDriveType(DriveType):
