--- pepper2/daemon/dbus/__init__.py
@@ -1,1 +1,0 @@
-"""D-Bus objects."""
