--- pepper2/daemon/dbus/controller.xml
@@ -1,21 +1,0 @@
-<!-- Controller DBus Configuration -->
-<!-- This XML is read by pydbus. -->
-<node>
-    <interface name='uk.org.j5.pepper2.Controller'>
-        <property name="daemon_status" type="s" access="read">
-          <annotation name="org.freedesktop.DBus.Property.EmitsChangedSignal" value="true"/>
-        </property>
-        <property name="version" type="s" access="read"></property>
-        <property name="usercode_drive" type="s" access="read"></property>
-        <property name="usercode_driver_name" type="s" access="read"></property>
-        <method name='get_drive_list'>
-            <arg type='as' name='drives' direction='out'/>
-        </method>
-        <method name='kill_usercode'>
-            <arg type='b' name='success' direction='out' />
-        </method>
-        <method name='start_usercode'>
-            <arg type='b' name='success' direction='out' />
-        </method>
-    </interface>
-</node>
